+++ iTestServer/client.h
@@ -49,7 +49,7 @@ public slots:
     void setName(QString); QString name();
     void setNumber(int); int number();
     void setSocket(QTcpSocket *); QTcpSocket * socket();
-    float score(); float maximumScore();
+    double score(); double maximumScore();
     bool isReady();
     QMap<QString, QuestionAnswer> * results();
     void setPassed(bool); bool passed();
@@ -68,8 +68,8 @@ signals:
 private:
     QString c_name;
     QTcpSocket * c_socket;
-    float c_score;
-    float c_maxscore;
+    double c_score;
+    double c_maxscore;
     bool c_ready;
     QMap<QString, QuestionAnswer> * c_results;
     bool c_passed;
